refactor operator into submodule, initial work for non-crawler operators
