+++ backend/btrixcloud/operator/__init__.py
@@ -1,0 +1,23 @@
+""" operators module """
+
+from .profiles import ProfileOperator
+from .bgjobs import BgJobOperator
+from .cronjobs import CronJobOperator
+
+operator_classes = [ProfileOperator, BgJobOperator, CronJobOperator]
+
+# ============================================================================
+def init_operator_api(app, *args):
+    """regsiters webhook handlers for metacontroller"""
+
+    operators = []
+    for cls in operator_classes:
+        oper = cls(*args)
+        oper.init_routes(app)
+        operators.append(oper)
+
+    @app.get("/healthz", include_in_schema=False)
+    async def healthz():
+        return {}
+
+    return operators
