+++ backend/btrixcloud/operator/crawls.py
@@ -1,12 +1,12 @@
 """ btrixjob operator (working for metacontroller) """
 
+from operator.profileoperator import ProfileOperator
+
 import asyncio
 import traceback
 import os
 from pprint import pprint
 from typing import Optional, DefaultDict, TYPE_CHECKING
-
-from collections import defaultdict
 
 import json
 from uuid import UUID
@@ -14,8 +14,6 @@ from fastapi import HTTPException
 
 import yaml
 import humanize
-
-from pydantic import BaseModel, Field
 
 from kubernetes.utils import parse_quantity
 from redis import asyncio as exceptions
@@ -53,207 +51,6 @@ if TYPE_CHECKING:
 else:
     CrawlConfigOps = CrawlOps = OrgOps = CollectionOps = Redis = object
     StorageOps = EventWebhookOps = UserManager = BackgroundJobOps = PageOps = object
-
-CMAP = "ConfigMap.v1"
-PVC = "PersistentVolumeClaim.v1"
-POD = "Pod.v1"
-
-BTRIX_API = "btrix.cloud/v1"
-CJS = f"CrawlJob.{BTRIX_API}"
-
-METRICS_API = "metrics.k8s.io/v1beta1"
-METRICS = f"PodMetrics.{METRICS_API}"
-
-DEFAULT_TTL = 30
-
-REDIS_TTL = 60
-
-# time in seconds before a crawl is deemed 'waiting' instead of 'starting'
-STARTING_TIME_SECS = 60
-
-# how often to update execution time seconds
-EXEC_TIME_UPDATE_SECS = 60
-
-
-# ============================================================================
-class MCBaseRequest(BaseModel):
-    """base metacontroller model, used for customize hook"""
-
-    parent: dict
-    controller: dict
-
-
-# ============================================================================
-class MCSyncData(MCBaseRequest):
-    """sync / finalize metacontroller model"""
-
-    children: dict
-    related: dict
-    finalizing: bool = False
-
-
-# ============================================================================
-class MCDecoratorSyncData(BaseModel):
-    """sync for decoratorcontroller model"""
-
-    object: dict
-    controller: dict
-
-    attachments: dict
-    related: dict
-    finalizing: bool = False
-
-
-# ============================================================================
-class CrawlSpec(BaseModel):
-    """spec from k8s CrawlJob object"""
-
-    id: str
-    cid: UUID
-    oid: UUID
-    scale: int = 1
-    storage: StorageRef
-    started: str
-    crawler_channel: str
-    stopping: bool = False
-    scheduled: bool = False
-    timeout: int = 0
-    max_crawl_size: int = 0
-
-
-# ============================================================================
-class PodResourcePercentage(BaseModel):
-    """Resource usage percentage ratios"""
-
-    memory: float = 0
-    cpu: float = 0
-    storage: float = 0
-
-
-# ============================================================================
-class PodResources(BaseModel):
-    """Pod Resources"""
-
-    memory: int = 0
-    cpu: float = 0
-    storage: int = 0
-
-    def __init__(self, *a, **kw):
-        if "memory" in kw:
-            kw["memory"] = int(parse_quantity(kw["memory"]))
-        if "cpu" in kw:
-            kw["cpu"] = float(parse_quantity(kw["cpu"]))
-        if "storage" in kw:
-            kw["storage"] = int(parse_quantity(kw["storage"]))
-        super().__init__(*a, **kw)
-
-
-# ============================================================================
-class PodInfo(BaseModel):
-    """Aggregate pod status info held in CrawlJob"""
-
-    exitTime: Optional[str] = None
-    exitCode: Optional[int] = None
-    isNewExit: Optional[bool] = Field(default=None, exclude=True)
-    reason: Optional[str] = None
-
-    allocated: PodResources = PodResources()
-    used: PodResources = PodResources()
-
-    newCpu: Optional[int] = None
-    newMemory: Optional[int] = None
-
-    def dict(self, *a, **kw):
-        res = super().dict(*a, **kw)
-        percent = {
-            "memory": self.get_percent_memory(),
-            "cpu": self.get_percent_cpu(),
-            "storage": self.get_percent_storage(),
-        }
-        res["percent"] = percent
-        return res
-
-    def get_percent_memory(self) -> float:
-        """compute percent memory used"""
-        return (
-            float(self.used.memory) / float(self.allocated.memory)
-            if self.allocated.memory
-            else 0
-        )
-
-    def get_percent_cpu(self) -> float:
-        """compute percent cpu used"""
-        return (
-            float(self.used.cpu) / float(self.allocated.cpu)
-            if self.allocated.cpu
-            else 0
-        )
-
-    def get_percent_storage(self) -> float:
-        """compute percent storage used"""
-        return (
-            float(self.used.storage) / float(self.allocated.storage)
-            if self.allocated.storage
-            else 0
-        )
-
-    def should_restart_pod(self):
-        """return true if pod should be restarted"""
-        if self.newMemory and self.newMemory != self.allocated.memory:
-            return True
-
-        if self.newCpu and self.newCpu != self.allocated.cpu:
-            return True
-
-        return False
-
-
-# ============================================================================
-class CrawlStatus(BaseModel):
-    """status from k8s CrawlJob object"""
-
-    state: str = "starting"
-    pagesFound: int = 0
-    pagesDone: int = 0
-    size: int = 0
-    # human readable size string
-    sizeHuman: str = ""
-    scale: int = 1
-    filesAdded: int = 0
-    filesAddedSize: int = 0
-    finished: Optional[str] = None
-    stopping: bool = False
-    stopReason: Optional[str] = None
-    initRedis: bool = False
-    crawlerImage: Optional[str] = None
-    lastActiveTime: str = ""
-    podStatus: Optional[DefaultDict[str, PodInfo]] = defaultdict(
-        lambda: PodInfo()  # pylint: disable=unnecessary-lambda
-    )
-    # placeholder for pydantic 2.0 -- will require this version
-    # podStatus: Optional[
-    #    DefaultDict[str, Annotated[PodInfo, Field(default_factory=PodInfo)]]
-    # ]
-    restartTime: Optional[str]
-    canceled: bool = False
-
-    # updated on pod exits and at regular interval
-    # Crawl Execution Time -- time all crawler pods have been running
-    # used to track resource usage and enforce execution minutes limit
-    crawlExecTime: int = 0
-
-    # Elapsed Exec Time -- time crawl has been running in at least one pod
-    # used for crawl timeouts
-    elapsedCrawlTime: int = 0
-
-    # last exec time update
-    lastUpdatedTime: str = ""
-
-    # any pods exited
-    anyCrawlPodNewExit: Optional[bool] = Field(default=False, exclude=True)
-
-    # don't include in status, use by metacontroller
-    resync_after: Optional[int] = Field(default=None, exclude=True)
 
 
 # ============================================================================
@@ -347,40 +144,6 @@ class BtrixOperator(K8sAPI):
         """perform any async init here"""
         self._has_pod_metrics = await self.is_pod_metrics_available()
         print("Pod Metrics Available:", self._has_pod_metrics)
-
-    async def sync_profile_browsers(self, data: MCSyncData):
-        """sync profile browsers"""
-        spec = data.parent.get("spec", {})
-
-        expire_time = from_k8s_date(spec.get("expireTime"))
-        browserid = spec.get("id")
-
-        if dt_now() >= expire_time:
-            self.run_task(self.delete_profile_browser(browserid))
-            return {"status": {}, "children": []}
-
-        params = {}
-        params.update(self.shared_params)
-        params["id"] = browserid
-        params["userid"] = spec.get("userid", "")
-
-        oid = spec.get("oid")
-        storage = StorageRef(spec.get("storageName"))
-
-        storage_path = storage.get_storage_extra_path(oid)
-        storage_secret = storage.get_storage_secret_name(oid)
-
-        params["storage_path"] = storage_path
-        params["storage_secret"] = storage_secret
-        params["profile_filename"] = spec.get("profileFilename", "")
-        params["crawler_image"] = spec["crawlerImage"]
-
-        params["url"] = spec.get("startUrl", "about:blank")
-        params["vnc_password"] = spec.get("vncPassword")
-
-        children = self.load_from_yaml("profilebrowser.yaml", params)
-
-        return {"status": {}, "children": children}
 
     # pylint: disable=too-many-return-statements, invalid-name
     async def sync_crawls(self, data: MCSyncData):
@@ -1583,192 +1346,3 @@ class BtrixOperator(K8sAPI):
         finally:
             if redis:
                 await redis.close()
-
-    def get_cronjob_crawl_related(self, data: MCBaseRequest):
-        """return configmap related to crawl"""
-        labels = data.parent.get("metadata", {}).get("labels", {})
-        cid = labels.get("btrix.crawlconfig")
-        return {
-            "relatedResources": [
-                {
-                    "apiVersion": "v1",
-                    "resource": "configmaps",
-                    "labelSelector": {"matchLabels": {"btrix.crawlconfig": cid}},
-                }
-            ]
-        }
-
-    async def sync_cronjob_crawl(self, data: MCDecoratorSyncData):
-        """create crawljobs from a job object spawned by cronjob"""
-
-        metadata = data.object["metadata"]
-        labels = metadata.get("labels", {})
-        cid = labels.get("btrix.crawlconfig")
-
-        name = metadata.get("name")
-        crawl_id = name
-
-        actual_state, finished = await self.crawl_ops.get_crawl_state(crawl_id)
-        if finished:
-            status = None
-            # mark job as completed
-            if not data.object["status"].get("succeeded"):
-                print("Cron Job Complete!", finished)
-                status = {
-                    "succeeded": 1,
-                    "startTime": metadata.get("creationTimestamp"),
-                    "completionTime": to_k8s_date(finished),
-                }
-
-            return {
-                "attachments": [],
-                "annotations": {"finished": finished},
-                "status": status,
-            }
-
-        configmap = data.related[CMAP][f"crawl-config-{cid}"]["data"]
-
-        oid = configmap.get("ORG_ID")
-        userid = configmap.get("USER_ID")
-
-        crawljobs = data.attachments[CJS]
-
-        org = await self.org_ops.get_org_by_id(UUID(oid))
-
-        warc_prefix = None
-
-        if not actual_state:
-            # cronjob doesn't exist yet
-            crawlconfig = await self.crawl_config_ops.get_crawl_config(
-                UUID(cid), UUID(oid)
-            )
-            if not crawlconfig:
-                print(
-                    f"error: no crawlconfig {cid}. skipping scheduled job. old cronjob left over?"
-                )
-                return {"attachments": []}
-
-            # db create
-            user = await self.user_ops.get_by_id(UUID(userid))
-            if not user:
-                print(f"error: missing user for id {userid}")
-                return {"attachments": []}
-
-            warc_prefix = self.crawl_config_ops.get_warc_prefix(org, crawlconfig)
-
-            await self.crawl_config_ops.add_new_crawl(
-                crawl_id,
-                crawlconfig,
-                user,
-                manual=False,
-            )
-            print("Scheduled Crawl Created: " + crawl_id)
-
-        crawl_id, crawljob = self.new_crawl_job_yaml(
-            cid,
-            userid=userid,
-            oid=oid,
-            storage=org.storage,
-            crawler_channel=configmap.get("CRAWLER_CHANNEL", "default"),
-            scale=int(configmap.get("INITIAL_SCALE", 1)),
-            crawl_timeout=int(configmap.get("CRAWL_TIMEOUT", 0)),
-            max_crawl_size=int(configmap.get("MAX_CRAWL_SIZE", "0")),
-            manual=False,
-            crawl_id=crawl_id,
-            warc_prefix=warc_prefix,
-        )
-
-        attachments = list(yaml.safe_load_all(crawljob))
-
-        if crawl_id in crawljobs:
-            attachments[0]["status"] = crawljobs[CJS][crawl_id]["status"]
-
-        return {
-            "attachments": attachments,
-        }
-
-    async def finalize_background_job(self, data: MCDecoratorSyncData) -> dict:
-        """handle finished background job"""
-
-        metadata = data.object["metadata"]
-        labels: dict[str, str] = metadata.get("labels", {})
-        oid: str = labels.get("btrix.org") or ""
-        job_type: str = labels.get("job_type") or ""
-        job_id: str = metadata.get("name")
-
-        status = data.object["status"]
-        success = status.get("succeeded") == 1
-        completion_time = status.get("completionTime")
-
-        finalized = True
-
-        finished = from_k8s_date(completion_time) if completion_time else dt_now()
-
-        try:
-            await self.background_job_ops.job_finished(
-                job_id, job_type, UUID(oid), success=success, finished=finished
-            )
-            # print(
-            #    f"{job_type} background job completed: success: {success}, {job_id}",
-            #    flush=True,
-            # )
-
-        # pylint: disable=broad-except
-        except Exception:
-            print("Update Background Job Error", flush=True)
-            traceback.print_exc()
-
-        return {"attachments": [], "finalized": finalized}
-
-    def run_task(self, func):
-        """add bg tasks to set to avoid premature garbage collection"""
-        task = asyncio.create_task(func)
-        self.bg_tasks.add(task)
-        task.add_done_callback(self.bg_tasks.discard)
-
-
-# ============================================================================
-def init_operator_api(app, *args):
-    """regsiters webhook handlers for metacontroller"""
-
-    oper = BtrixOperator(*args)
-
-    @app.post("/op/crawls/sync")
-    async def mc_sync_crawls(data: MCSyncData):
-        return await oper.sync_crawls(data)
-
-    # reuse sync path, but distinct endpoint for better logging
-    @app.post("/op/crawls/finalize")
-    async def mc_sync_finalize(data: MCSyncData):
-        return await oper.sync_crawls(data)
-
-    @app.post("/op/crawls/customize")
-    async def mc_related(data: MCBaseRequest):
-        return oper.get_related(data)
-
-    @app.post("/op/profilebrowsers/sync")
-    async def mc_sync_profile_browsers(data: MCSyncData):
-        return await oper.sync_profile_browsers(data)
-
-    @app.post("/op/cronjob/sync")
-    async def mc_sync_cronjob_crawls(data: MCDecoratorSyncData):
-        return await oper.sync_cronjob_crawl(data)
-
-    @app.post("/op/cronjob/customize")
-    async def mc_cronjob_related(data: MCBaseRequest):
-        return oper.get_cronjob_crawl_related(data)
-
-    # nop, but needed for metacontroller
-    @app.post("/op/backgroundjob/sync")
-    async def mc_sync_background_jobs():
-        return {"attachments": []}
-
-    @app.post("/op/backgroundjob/finalize")
-    async def mc_finalize_background_jobs(data: MCDecoratorSyncData):
-        return await oper.finalize_background_job(data)
-
-    @app.get("/healthz", include_in_schema=False)
-    async def healthz():
-        return {}
-
-    return oper
